refactor reading a member by name, enabling dotty java enums
