+++ src/compiler/scala/tools/nsc/tasty/bridge/NameOps.scala
@@ -1,6 +1,6 @@
 package scala.tools.nsc.tasty.bridge
 
-import scala.tools.nsc.tasty.Names.TastyName
+import scala.tools.nsc.tasty.TastyName
 import scala.tools.nsc.tasty.TastyUniverse
 
 trait NameOps { self: TastyUniverse =>
@@ -9,12 +9,12 @@ trait NameOps { self: TastyUniverse =>
   def isConstructorName(name: Name) = symbolTable.nme.isConstructorName(name)
 
   def encodeAsTermName(tastyName: TastyName): TermName = tastyName match {
-    case Empty          => termNames.EMPTY
-    case Constructor    => nme.CONSTRUCTOR
-    case EmptyPkg       => nme.EMPTY_PACKAGE_NAME
-    case RootClass      => nme.ROOT
-    case WildcardName() => nme.WILDCARD
-    case name           => mkTermName(name.encoded)
+    case Empty                                    => termNames.EMPTY
+    case Constructor | SignedName(Constructor, _) => nme.CONSTRUCTOR
+    case EmptyPkg                                 => nme.EMPTY_PACKAGE_NAME
+    case RootClass                                => nme.ROOT
+    case WildcardName()                           => nme.WILDCARD
+    case name                                     => mkTermName(name.encoded)
   }
 
   def encodeTastyName(tastyName: TastyName, isTerm: Boolean): Name = {
