+++ src/compiler/scala/tools/nsc/tasty/bridge/SymbolOps.scala
@@ -5,6 +5,8 @@ import scala.tools.nsc.tasty.SafeEq
 import scala.tools.nsc.tasty.TastyUniverse
 import scala.tools.nsc.tasty.Signature
 import scala.tools.nsc.tasty.Signature.MethodSignature
+import scala.tools.nsc.tasty.TastyName
+import scala.tools.nsc.tasty.Signature.NotAMethod
 
 trait SymbolOps { self: TastyUniverse =>
   import Contexts.Context
@@ -33,11 +35,34 @@ trait SymbolOps { self: TastyUniverse =>
     }
   }
 
-  def selectSymFromSig(space: Type, name: Name, sig: Signature[Type])(implicit ctx: Context): Either[String, (Int, Symbol)] = {
+  def constructorOfType(space: Type): Either[String, Symbol] =
+    space.member(nme.CONSTRUCTOR).asTerm.alternatives.find(_.isInstanceOf[MethodSymbol]).toRight(
+      s"${space.typeSymbol} has no constructor"
+    )
+
+  def namedMemberOfType(space: Type, tname: TastyName, selectingTerm: Boolean)(implicit ctx: Context): Either[String, Symbol] = {
+    val selector = encodeTastyName(tname, selectingTerm)
+    tname.signature match {
+      case NotAMethod => Right(memberOfSpace(space, selector, tname.isModuleName))
+      case sig        => signedMemberOfSpace(space, selector, sig.map(erasedNameToErasedType))
+    }
+  }
+
+  private def memberOfSpace(space: Type, name: Name, isModuleName: Boolean): Symbol = {
+    // TODO [tasty]: dotty uses accessibleDenot which asserts that `fetched.isAccessibleFrom(pre)`,
+    //    or else filters for non private.
+    // There should be an investigation to see what code makes that false, and what is an equivalent check.
+    def lookInTypeCtor = space.typeConstructor.typeParams.filter(_.name == name).headOption.getOrElse(noSymbol)
+    val fetched = space.member(name)
+    val corrected = if (name.isTypeName) fetched.orElse(lookInTypeCtor) else fetched
+    if (isModuleName) corrected.linkedClassOfClass else corrected
+  }
+
+  private def signedMemberOfSpace(space: Type, name: Name, sig: Signature[Type])(implicit ctx: Context): Either[String, Symbol] = {
     ctx.log(s"""looking for overload member[$space]("$name") @@ ${sig.show}""")
     val MethodSignature(args, ret) = sig
     val member = space.member(name)
-    val (tyParamCount, argsSyms) = {
+    val (tyParamCount, argTpes) = {
       val (tyParamCounts, params) = args.partitionMap(identity)
       if (tyParamCounts.length > 1) {
         reporter.error(noPosition, s"Multiple type parameter lists on signature ${sig.show} for $member.")
@@ -48,17 +73,19 @@ trait SymbolOps { self: TastyUniverse =>
       case sym: MethodSymbol =>
         val params = sym.paramss.flatten
         sym.returnType.erasure =:= ret &&
-        params.length === argsSyms.length &&
-        ((name === nme.CONSTRUCTOR && tyParamCount === member.owner.typeParams.length)
+        params.length === argTpes.length &&
+        (name === nme.CONSTRUCTOR && tyParamCount === member.owner.typeParams.length
           || tyParamCount === sym.typeParams.length) &&
-        params.zip(argsSyms).forall { case (param, tpe) => param.tpe.erasure =:= tpe }
+        params.zip(argTpes).forall { case (param, tpe) => param.tpe.erasure =:= tpe } && {
+          ctx.log(s"selected ${showSym(sym)} : ${sym.tpe}")
+          true
+        }
       case _ =>
         ctx.log(s"""member[$space]("$name") ${showSym(sym)} is not a method""")
         false
     }
-    member.asTerm.alternatives.find(compareSym)
-      .map(tyParamCount -> _)
-      .toRight(s"No matching overload of $space.$name with signature ${sig.show}")
+    member.asTerm.alternatives.find(compareSym).toRight(
+      s"No matching overload of $space.$name with signature ${sig.show}")
   }
 
   def showSym(sym: Symbol): String = s"Symbol($sym, #${sym.hashCode})"
