+++ test/tasty/run/src-2/tastytest/TestJColour.scala
@@ -1,7 +1,5 @@
 package tastytest
 
-/** Suspended as looking for a non-existing ctor with sig (1)java.lang.Enum causes a compiler crash
- */
 object TestJColour extends Suite("TestJColour") {
 
   test(assert(JColour.Red == JColour.Red))
